moved java classes from server/dhcp-sar to services/dhcp, removed jboss specific code
