+++ services/dhcp/pom.xml
@@ -12,5 +12,37 @@
 
     <artifactId>manager-service-dhcp</artifactId>
 
+    <dependencies>
     
+        <!--
+        OUR COMMON API THAT ALL SERVICES SHARE
+        -->
+        <dependency>
+            <groupId>org.openthinclient</groupId>
+            <artifactId>manager-service-common</artifactId>
+            <version>2.0.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>log4j</groupId>
+            <artifactId>log4j</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openthinclient</groupId>
+            <artifactId>manager-common</artifactId>
+            <version>2.0.0-SNAPSHOT</version>
+        </dependency>
+                        
+        <!--
+            REQUIRED LIBRARIES FOR THE APACHE DS
+        -->
+        <dependency>
+
+            <groupId>org.openthinclient.3rd-party.apacheds</groupId>
+            <artifactId>apacheds-protocol-dhcp</artifactId>
+            <version>1.0.2-SNAPSHOT</version>
+        </dependency>
+
+    </dependencies>    
 </project>
