+++ services/dhcp/src/main/java/org/openthinclient/service/dhcp/AbstractPXEService.java
@@ -18,7 +18,7 @@
  * this program; if not, write to the Free Software Foundation, Inc., 59 Temple
  * Place - Suite 330, Boston, MA 02111-1307, USA.
  ******************************************************************************/
-package org.openthinclient.dhcp;
+package org.openthinclient.service.dhcp;
 
 import java.io.IOException;
 import java.net.InetAddress;
