+++ services/dhcp/src/main/java/org/openthinclient/service/dhcp/DhcpServiceConfiguration.java
@@ -1,0 +1,9 @@
+package org.openthinclient.service.dhcp;
+import org.openthinclient.service.common.home.Configuration;
+import org.openthinclient.service.common.home.ConfigurationFile;
+
+
+public class DhcpServiceConfiguration implements Configuration {
+
+
+}
