+++ services/dhcp/src/main/java/org/openthinclient/service/dhcp/SingleHomedPXEService.java
@@ -1,11 +1,10 @@
-package org.openthinclient.dhcp;
+package org.openthinclient.service.dhcp;
 
 import java.io.IOException;
 import java.net.Inet4Address;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.NetworkInterface;
-import java.net.SocketException;
 import java.util.Enumeration;
 
 import org.apache.mina.common.IoAcceptor;
@@ -21,24 +20,23 @@ import org.openthinclient.ldap.DirectoryException;
  * 
  * @author levigo
  */
-public class SingleHomedBroadcastPXEService extends BasePXEService {
+public class SingleHomedPXEService extends BasePXEService {
 	private InetAddress serverAddress;
 
-	public SingleHomedBroadcastPXEService() throws DirectoryException {
+	public SingleHomedPXEService() throws DirectoryException {
 		super();
 	}
 
 	/*
 	 * (non-Javadoc)
 	 * 
-	 * @see
-	 * org.openthinclient.dhcp.BasePXEService#determineServerAddress(java.net.
-	 * InetSocketAddress)
+	 * @see org.openthinclient.dhcp.BasePXEService#determineServerAddress(java.net.InetSocketAddress)
 	 */
 	@Override
 	protected InetSocketAddress determineServerAddress(
 			InetSocketAddress localAddress) {
-		return new InetSocketAddress(serverAddress, 67);
+		// since we bound individually, we can use the local address.
+		return localAddress;
 	}
 
 	@Override
@@ -46,35 +44,16 @@ public class SingleHomedBroadcastPXEService extends BasePXEService {
 			IoServiceConfig config) throws IOException {
 		logger
 				.warn("-------------------------------------------------------------");
-		logger.warn("  Using SingleHomedBroadcastPXEService implementation. ");
+		logger.warn("  Using SingleHomedPXEService implementation. ");
 		logger
 				.warn("  This type of service might be problematic on multi-homed systems.");
 		logger.warn("  (for more details, see log messages with level INFO)");
 		logger.info("");
 
-		serverAddress = getConfiguredServerAddress();
+		InetAddress localAddress = getConfiguredLocalAddress();
 
-		final InetSocketAddress listenPort = new InetSocketAddress(67);
-		acceptor.bind(listenPort, handler, config);
-		logger.info("  Binding on " + listenPort);
-
-		final InetSocketAddress pxePort = new InetSocketAddress(4011);
-		acceptor.bind(pxePort, handler, config);
-		logger.info("  Binding on " + pxePort);
-
-		logger
-				.warn("-------------------------------------------------------------");
-	}
-
-	/**
-	 * Determine a configured static server address.
-	 * 
-	 * @return
-	 */
-	private InetAddress getConfiguredServerAddress() {
-		try {
+		if (null == localAddress)
 			// auto-determine local address
-			InetAddress ifAddress = null;
 			outer : for (final Enumeration i = NetworkInterface
 					.getNetworkInterfaces(); i.hasMoreElements();) {
 				final NetworkInterface nif = (NetworkInterface) i.nextElement();
@@ -84,20 +63,36 @@ public class SingleHomedBroadcastPXEService extends BasePXEService {
 							.hasMoreElements();) {
 						final InetAddress address = (InetAddress) j.nextElement();
 						if (address instanceof Inet4Address && !address.isLoopbackAddress()) {
-							ifAddress = address;
-							logger.info("  Using address " + ifAddress + " as source IP");
+							localAddress = address;
 							break outer;
 						}
 					}
 			}
-			if (null == ifAddress) {
-				logger.error("  No non loopback InterfaceAddress found at all");
-				return null;
-			}
-			return ifAddress;
-		} catch (final SocketException e) {
-			logger.error("  Can't determine network interface");
+
+		if (null == localAddress)
+			logger
+					.warn("  Could not determine ANY local address to use. PXE service will be unavailable");
+		else {
+			serverAddress = localAddress;
+			acceptor.bind(new InetSocketAddress(serverAddress, 67), handler, config);
+			logger.info("  Binding on " + serverAddress);
+
+			final InetSocketAddress pxePort = new InetSocketAddress(localAddress,
+					4011);
+			acceptor.bind(pxePort, handler, config);
+			logger.info("  Binding on " + pxePort);
 		}
-		return null;
+
+		logger
+				.warn("-------------------------------------------------------------");
+	}
+
+	/**
+	 * Determine a configured static local address.
+	 * 
+	 * @return
+	 */
+	private InetAddress getConfiguredLocalAddress() {
+		return serverAddress;
 	}
 }
